+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.3
+- PR #5 - Added ability to add read and write transit handlers to the default set.
+- Fixed dependencies from provided to test for spec
+
+# 0.3.2
+- Added the ability to define the host for `make-channel-client`.
+
 # 0.3.1
 - Fix initialization timing issue where websockets messages were being received before multimethods were installed.